requirements.txt
psutil==5.6.0
PyQt5==5.12
PyYAML==3.13
